backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
requests==2.31.0
psycopg2-binary>=2.9.10
gunicorn==21.2.0
django-cors-headers==4.3.0
python-decouple==3.8
dj-database-url==2.1.0
